+++ intcode/src/builder.rs
@@ -12,7 +12,7 @@ impl Builder {
         }
     }
 
-    pub fn program(&mut self, program: &Vec<i32>) -> &mut Self {
+    pub fn program(&mut self, program: &Vec<i128>) -> &mut Self {
         self._program = Some(program.clone());
         self
     }
@@ -23,24 +23,24 @@ impl Builder {
                 .split(",")
                 .filter(|s| s.len() > 0)
                 .map(|s| {
-                    s.parse::<i32>()
+                    s.parse::<i128>()
                         .expect(&format!("invalid instruction: '{}'", s))
                 })
                 .collect(),
         )
     }
 
-    pub fn init_hook(&mut self, hook: &Vec<i32>) -> &mut Self {
+    pub fn init_hook(&mut self, hook: &Vec<i128>) -> &mut Self {
         self._init_hook = Some(hook.clone());
         self
     }
 
-    pub fn exit_hook(&mut self, hook: &Vec<i32>) -> &mut Self {
+    pub fn exit_hook(&mut self, hook: &Vec<i128>) -> &mut Self {
         self._exit_hook = Some(hook.clone());
         self
     }
 
-    pub fn run(&self, input: Receiver<i32>, output: Sender<i32>) {
+    pub fn run(&self, input: Receiver<i128>, output: Sender<i128>) {
         let mut computer = IntcodeComputer::new(
             self._program.as_ref().expect("No program set!"),
             self._init_hook.clone(),
@@ -54,7 +54,7 @@ impl Builder {
         });
     }
 
-    pub fn run_noninteractive(&self, input: &mut dyn Iterator<Item = i32>) -> Vec<i32> {
+    pub fn run_noninteractive(&self, input: &mut dyn Iterator<Item = i128>) -> Vec<i128> {
         let (in_tx, in_rx) = channel();
         let (out_tx, out_rx) = channel();
