+++ intcode/tests/intcode.rs
@@ -1,14 +1,14 @@
 use intcode::*;
 use std::iter::{empty, once, repeat};
 
-fn get_value_at_0(program: &Vec<i32>, input: &mut dyn Iterator<Item = i32>) -> i32 {
+fn get_value_at_0(program: &Vec<i128>, input: &mut dyn Iterator<Item = i128>) -> i128 {
   let output = Builder::new()
     .program(&program)
     .exit_hook(&vec![4, 0, 99])
     .run_noninteractive(input);
   *output.last().unwrap()
 }
-fn get_last_output(program: &Vec<i32>, input: &mut dyn Iterator<Item = i32>) -> i32 {
+fn get_last_output(program: &Vec<i128>, input: &mut dyn Iterator<Item = i128>) -> i128 {
   let output = Builder::new().program(&program).run_noninteractive(input);
   *output.last().unwrap()
 }
